Fix bug: Unload Sources doesn't close all characters first Issue#: CODE-1565
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@17512 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/io/IOHandler.java
@@ -164,7 +164,7 @@ public abstract class IOHandler
 				{
 					bakFile = new File(filename + BAK_PREFIX);
 				}
-				if (bakFile.exists())
+				if (bakFile.exists() && outFile.exists() && outFile.length() > 0)
 				{
 					bakFile.delete();
 				}
